Convert launch script to VisualizationTool class
